+++ src/Sentry/HttpHeadersExtensions.cs
@@ -1,3 +1,6 @@
+using Sentry.Internal;
+using Sentry.Internal.Extensions;
+
 namespace Sentry;
 
 internal static class HttpHeadersExtensions
@@ -6,4 +9,14 @@ internal static class HttpHeadersExtensions
         headers.TryGetValues("Cookie", out var values)
             ? string.Join("; ", values)
             : string.Empty;
+
+    internal static RedactedHeaders? Redact(this Dictionary<string, string?>? headers)
+    {
+        var items = headers?.WhereNotNullValue();
+        if (items is null || !items.Any())
+        {
+            return null;
+        }
+        return headers!;
+    }
 }